[Refactor](exec) remove the unless file and refactor the agg code
