+++ be/src/vec/utils/util.hpp
@@ -148,12 +148,7 @@ public:
     }
 
     static bool all_arguments_are_constant(const Block& block, const ColumnNumbers& args) {
-        for (const auto& arg : args) {
-            if (!is_column_const(*block.get_by_position(arg).column)) {
-                return false;
-            }
-        }
-        return true;
+        return std::ranges::all_of(block, [](auto& data) { return is_column_const(*data.column); });
     }
 };
